Abstract a function
