+++ booking-app/scripts/copyCollection.js
@@ -2,6 +2,11 @@
 require("dotenv").config({ path: ".env.local" });
 const admin = require("firebase-admin");
 const fs = require("fs");
+const {
+  TENANT_SCHEMA_COLLECTION,
+  backupTenantSchemaCollection,
+  createBackupDocId,
+} = require("./tenantSchemaBackup");
 
 // Database names for different environments - these can be overridden
 const DATABASES = {
@@ -10,8 +15,6 @@ const DATABASES = {
   production: "booking-app-prod",
 };
 const BACKUP_TYPE_COPY = "copy";
-const TENANT_SCHEMA_COLLECTION = "tenantSchema";
-const TENANT_SCHEMA_BACKUP_COLLECTION = "tenantSchemaBackup";
 
 const isPlainObject = (value) =>
   value !== null && typeof value === "object" && !Array.isArray(value);
@@ -70,6 +73,18 @@ const compareDocumentData = (before, after) => {
     updatedValues,
   };
 };
+
+const backupTenantSchemaAsDocuments = (
+  targetDb,
+  databaseName,
+  dryRun = false
+) =>
+  backupTenantSchemaCollection(
+    targetDb,
+    databaseName,
+    BACKUP_TYPE_COPY,
+    dryRun
+  );
 
 // Parse command line arguments
 const parseArgs = () => {
@@ -232,84 +247,6 @@ const testDatabaseConnection = async (db, databaseName) => {
       console.log(`❌ Error connecting to ${databaseName}:`, error.message);
     }
     return false;
-  }
-};
-
-const createBackupDocId = (tenantId, backupType) => {
-  const timestamp = new Date()
-    .toISOString()
-    .replace(/[:.]/g, "-")
-    .replace("T", "_")
-    .replace("Z", "");
-  return `${tenantId}-backup-${backupType}-${timestamp}`;
-};
-
-const backupTenantSchemaAsDocuments = async (
-  targetDb,
-  databaseName,
-  dryRun = false
-) => {
-  try {
-    const targetSnapshot = await targetDb
-      .collection(TENANT_SCHEMA_COLLECTION)
-      .get();
-
-    if (targetSnapshot.empty) {
-      console.log(`ℹ️ No existing ${TENANT_SCHEMA_COLLECTION} documents to back up`);
-      return { success: true, copied: 0, errors: [] };
-    }
-
-    if (dryRun) {
-      console.log(
-        `\n🔍 [DRY RUN] Would back up ${targetSnapshot.size} ${TENANT_SCHEMA_COLLECTION} documents in ${databaseName} to ${TENANT_SCHEMA_BACKUP_COLLECTION}`
-      );
-      targetSnapshot.docs.forEach((doc) => {
-        const backupDocId = createBackupDocId(doc.id, BACKUP_TYPE_COPY);
-        console.log(
-          `  📄 Would create backup document: ${TENANT_SCHEMA_BACKUP_COLLECTION}/${backupDocId}`
-        );
-      });
-      return { success: true, copied: targetSnapshot.size, errors: [] };
-    }
-
-    console.log(
-      `\n📦 Backing up ${targetSnapshot.size} ${TENANT_SCHEMA_COLLECTION} documents in ${databaseName} to ${TENANT_SCHEMA_BACKUP_COLLECTION}...`
-    );
-    let copiedCount = 0;
-    let currentBatch = targetDb.batch();
-    let operationsInBatch = 0;
-    const BATCH_SIZE = 500;
-
-    for (const doc of targetSnapshot.docs) {
-      const backupDocId = createBackupDocId(doc.id, BACKUP_TYPE_COPY);
-      const backupDocRef = targetDb
-        .collection(TENANT_SCHEMA_BACKUP_COLLECTION)
-        .doc(backupDocId);
-      currentBatch.set(backupDocRef, doc.data(), { merge: false });
-      operationsInBatch++;
-      copiedCount++;
-
-      if (operationsInBatch >= BATCH_SIZE) {
-        await currentBatch.commit();
-        currentBatch = targetDb.batch();
-        operationsInBatch = 0;
-      }
-    }
-
-    if (operationsInBatch > 0) {
-      await currentBatch.commit();
-    }
-
-    console.log(
-      `✅ Backed up ${copiedCount} ${TENANT_SCHEMA_COLLECTION} documents to ${TENANT_SCHEMA_BACKUP_COLLECTION} in ${databaseName}`
-    );
-    return { success: true, copied: copiedCount, errors: [] };
-  } catch (error) {
-    console.error(
-      `❌ Error backing up ${TENANT_SCHEMA_COLLECTION} to ${TENANT_SCHEMA_BACKUP_COLLECTION} in ${databaseName}:`,
-      error.message
-    );
-    return { success: false, copied: 0, errors: [error.message] };
   }
 };
 
@@ -554,8 +491,8 @@ const main = async () => {
     // If we're updating tenantSchema, first backup current tenantSchema as backup documents
     const results = [];
     if (
-      options.targetCollection === "tenantSchema" &&
-      options.sourceCollection === "tenantSchema" &&
+      options.targetCollection === TENANT_SCHEMA_COLLECTION &&
+      options.sourceCollection === TENANT_SCHEMA_COLLECTION &&
       options.sourceDatabase !== options.targetDatabase
     ) {
       const backupResult = await backupTenantSchemaAsDocuments(
